+++ ReadADS1220_SPI/MovingAverageFilter.h
@@ -7,14 +7,14 @@
 class MovingAverageFilter {
     private:
         static const int WINDOW_SIZE = 10;
-        uint64_t window[WINDOW_SIZE];
-        uint64_t sortedWindow[WINDOW_SIZE];
+        uint32_t window[WINDOW_SIZE];
+        uint32_t sortedWindow[WINDOW_SIZE];
         int index;
         bool bufferFilled;
 
     public:
         MovingAverageFilter();
-        void addValue(uint64_t value);
+        void addValue(uint32_t value);
         float calculateFilteredValue();
         void clear();
 };
